updated ast pretty printing

+++ src/jet.c
@@ -91,7 +91,7 @@ static jet_ast* ast_generate(jet_da* tok_darray)
 static void print_source(const char* source, size_t src_len)
 {   
     puts("\n*****************************************************\n");
-    printf("source text loaded (len: %llu):\n\n%s", (unsigned long long)src_len, source);
+    printf("source text loaded (len: %zu):\n\n%s", src_len, source);
     puts("\n*****************************************************\n");
 }
